requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
azure-cosmos==4.5.1
python-dotenv==1.0.1
